knowledge Assistant/pyproject.toml
[project]
name = "knowledge-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.53.1",
    "wikipedia>=1.4.0",
]
